requirements.txt
selenium==4.10.0
pandas==2.0.3
PyYAML==6.0
requests==2.28.2
geopy==2.4.0
python-dotenv==1.0.0
numpy>=1.21
spacy>=3.0